Added Blocking Capability

+++ TreasureHunt/views.py
@@ -1,7 +1,7 @@
 from django.http import HttpResponse, HttpResponseRedirect
 from django.shortcuts import get_object_or_404, render
 from django.contrib.auth import logout
-from hunt.models import LoggedInUser
+from hunt.models import LoggedInUser, GameUser
 from django.core.exceptions import ObjectDoesNotExist
 from django.contrib.sessions.models import Session
 from django.contrib import messages
@@ -22,22 +22,27 @@ def leaderboard(request):
 
 def homepage_view(request):
     if request.user.is_authenticated:
-        msg = 'You are logged in!'
-        print("Back here")
+        # msg = 'You are logged in!'
+        # print("Back here")
         LoggedInUser.objects.get_or_create(user = request.user)
-        stored_session_key = request.user.logged_in_user.session_key
-        # if there is a stored_session_key  in our database and it is
-        # different from the current session, delete the stored_session_key
-        # session_key with from the Session table
-        print(stored_session_key)
-        if stored_session_key and stored_session_key != request.session.session_key:
-            print("Found Previous login")
-            Session.objects.get(session_key=stored_session_key).delete()
-        request.user.logged_in_user.session_key = request.session.session_key
-        request.user.logged_in_user.save()
-        social = request.user.social_auth.get(provider='facebook')
-        userid = social.uid
-        return HttpResponseRedirect('/hunt')
+        game_user = GameUser.objects.get(user = request.user)
+        if game_user.blocked:
+            messages.info(request,"Due to unfair play, your account has been blocked.")
+            logout_view(request)
+        else:        
+            stored_session_key = request.user.logged_in_user.session_key
+            # if there is a stored_session_key  in our database and it is
+            # different from the current session, delete the stored_session_key
+            # session_key with from the Session table
+            # print(stored_session_key)
+            if stored_session_key and stored_session_key != request.session.session_key:
+                # print("Found Previous login")
+                Session.objects.get(session_key=stored_session_key).delete()
+            request.user.logged_in_user.session_key = request.session.session_key
+            request.user.logged_in_user.save()
+            social = request.user.social_auth.get(provider='facebook')
+            userid = social.uid
+            return HttpResponseRedirect('/hunt')
     else:
         messages.info(request, 'You need to login first.')
         return HttpResponseRedirect('/')
